+++ src/server/data/account-provider.ts
@@ -1,10 +1,9 @@
-import { parseArangoConfig, QAccountProviderConfig } from "../config"
-import { Database } from "arangojs"
-import { createDatabase } from "./database-provider"
+import { QAccountProviderConfig } from "../config"
 import { RequestManager, HTTPTransport, Client } from "@open-rpc/client-js"
 
 export interface IAccountProvider {
     getBocs(addresses: string[]): Promise<Map<string, string>>
+    getMetas(addresses: string[]): Promise<Map<string, any>>
 }
 
 class NodeRpcProvider implements IAccountProvider {
@@ -33,34 +32,20 @@ class NodeRpcProvider implements IAccountProvider {
         }
         return resolved
     }
-}
 
-class ArangoProvider implements IAccountProvider {
-    private readonly database: Database
-    constructor(
-        public config: {
-            database: string
-            collection: string
-        },
-    ) {
-        this.database = createDatabase(parseArangoConfig(config.database))
-    }
-
-    async getBocs(addresses: string[]): Promise<Map<string, string>> {
+    async getMetas(addresses: string[]): Promise<Map<string, any>> {
         const resolved = new Map()
-        const cursor = await this.database.query(
-            `
-            FOR doc IN ${this.config.collection}
-            FILTER doc._key IN @addresses
-            RETURN { address: doc._key, boc: doc.boc }
-            `,
-            {
-                addresses,
-            },
-        )
-        const docs: { address: string; boc: string }[] = await cursor.all()
-        for (const doc of docs) {
-            resolved.set(doc.address, doc.boc)
+        // TODO: fetch bocs in parallel
+        for (const address of addresses) {
+            const result = await this.client.request({
+                method: "getAccountMeta",
+                params: {
+                    account: address,
+                },
+            })
+            if (result?.account_meta ?? "" !== "") {
+                resolved.set(address, result.account_meta)
+            }
         }
         return resolved
     }
@@ -74,9 +59,6 @@ export function createAccountProvider(
         return new NodeRpcProvider({
             endpoint: rpcEndpoint,
         })
-    }
-    if (config.arango && (config.arango.database ?? "" !== "")) {
-        return new ArangoProvider(config.arango)
     }
     return undefined
 }